Continue refactor, cleaning up
