+++ pyesasky/__init__.py
@@ -1,5 +1,5 @@
 from jupyter_server import serverapp
-from .pyesasky import ESASkyWidget  # noqa
+from pyesasky.widgets import ESASkyWidget  # noqa
 
 
 try:
@@ -7,7 +7,8 @@ try:
 except ImportError:
     # Fallback when using the package in dev mode without installing
     # in editable mode with pip. It is highly recommended to install
-    # the package from a stable release or in editable mode: https://pip.pypa.io/en/stable/topics/local-project-installs/#editable-installs
+    # the package from a stable release or in editable mode:
+    # https://pip.pypa.io/en/stable/topics/local-project-installs/#editable-installs
     import warnings
 
     warnings.warn("Importing 'pyesasky' outside a proper installation.")
